[cmyth] Fixed some warnings

+++ lib/cmyth/libcmyth/keyframe.c
@@ -117,7 +117,7 @@ cmyth_keyframe_string(cmyth_keyframe_t kf)
 	if (!kf) {
 		return NULL;
 	}
-	sprintf(pos, "%lld", kf->keyframe_pos);
+	sprintf(pos, "%lld", (long long)kf->keyframe_pos);
 	len += strlen(pos);
 	sprintf(key, "%ld", kf->keyframe_number);
 	len += strlen(key);
